backend/requirements.txt
asgiref==3.11.0
attrs==25.4.0
Django==6.0.1
django-filter==25.2
django-redis==6.0.0
django-seed==0.3.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
Faker==40.4.0
inflection==0.5.1
iniconfig==2.3.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
Markdown==3.10
packaging==26.0
pluggy==1.6.0
psycopg2==2.9.11
Pygments==2.19.2
PyJWT==2.10.1
pytest==9.0.2
python-decouple==3.8
PyYAML==6.0.3
redis==7.1.0
referencing==0.37.0
rest-framework-simplejwt==0.0.2
rpds-py==0.30.0
sqlparse==0.5.5
toposort==1.10
typing_extensions==4.15.0
uritemplate==4.2.0
